added function to seed.js to check for data then seed

+++ server/seeders/seed.js
@@ -3,17 +3,30 @@ const { Profile, Product } = require('../models');
 const profileSeeds = require('./profileSeeds.json');
 const productSeeds = require('./productSeeds.json')
 
-db.once('open', async () => {
-  try {
-    await Profile.deleteMany({});
-    await Profile.create(profileSeeds);
+// seed the databse function if no data exists
+function seedAll() {
+  db.once('open', async () => {
 
-    await Product.deleteMany({});
-    await Product.insertMany(productSeeds);
+    const dbData = await Profile.find({})
 
-    console.log('all done!');
-    process.exit(0);
-  } catch (err) {
-    throw err;
-  }
-});
+    if (dbData) {
+      console.log("data has already been seeded")
+    } else {
+      try {
+        await Profile.deleteMany({});
+        await Profile.create(profileSeeds);
+
+        await Product.deleteMany({});
+        await Product.insertMany(productSeeds);
+
+        console.log('all done!');
+        process.exit(0);
+      } catch (err) {
+        throw err;
+      }
+    }
+  });
+}
+
+module.exports = seedAll
+seedAll()
